flask-app/requirements.txt
Flask==3.0.3
mysql-connector-python==8.1.0
gunicorn==21.2.0